Implemented basic scanner and tests.
